feat: move fixtures to conftest for cleaner test file
